# Improve perf of _block_bucketize_sparse_features_cuda_kernel2 - for pooled emb